Legger til ledetekst på innhold i historikk
