+++ packages/deltaker-flate-common/index.ts
@@ -4,7 +4,7 @@ export * from './components/DeltakerStatusInfoTekst'
 export * from './components/UtkastHeader'
 export * from './components/HvaDelesMedArrangor'
 export * from './components/FeilregistrertInfo'
-export * from './components/DeltakelseInnholdListe'
+export * from './components/DeltakelseInnholdPanel.tsx'
 export * from './components/EndringTypeIkon'
 export * from './components/forslag/ForlengDeltakelseForslagDetaljer'
 export * from './components/forslag/ForslagDetaljer'
